refactor(nav): move helpers functions outside of test suite
- Move all extension functions in `NavigationTestHelpers` object
